+++ src/exec/mod.rs
@@ -6,12 +6,13 @@ use codespan_reporting::diagnostic::Diagnostic;
 use eyre::{bail, Result};
 
 use crate::ast::{
-    self, BinOp, Block, BlockType, Expr, Function, Item, Program, RawExpr, Span, Spanned, Stmt,
-    UnaryOp,
+    self, Block, BlockType, Expr, Function, Item, Program, RawExpr, Span, Spanned, Stmt, UnaryOp,
 };
 use crate::diagnostics::RtError;
 use crate::env::{Env, Scope};
-use crate::value::{BigValue, HeapKey, Value, ValueDbg};
+use crate::value::{BigValue, Value, ValueDbg};
+
+mod binop;
 
 // Err -> Exit err due to type error/rt error
 // Ok(false) -> Exit sucess
@@ -61,44 +62,6 @@ macro_rules! get {
         match $e {
             BlockEvalResult::LocalRet(x) => x,
             x @ BlockEvalResult::FnRet(_) => return Ok(x),
-        }
-    };
-}
-
-macro_rules! binop_match {
-    (
-        $bindings:expr,
-        ($l_span:expr, $o_span:expr, $r_span:expr, $env:expr),
-        // Integer math ops
-        { $($math_op_name:path => $math_op:tt),*$(,)? },
-        // Equality
-        // { $($eq_type:path),*$(,)? },
-        // Comparison
-        { $($comarison_name:path => $comparison_op:tt),*$(,)? },
-        // Misc user stuff
-        { $($user_lhs:pat => $user_rhs:expr),*$(,)? },
-    ) => {
-        // TODO: decide if this is a good idea
-        #[allow(clippy::float_cmp)]
-        match $bindings {
-            $(
-                // TODO: Should we coerce int to float in 1 + 2.0
-                ($math_op_name, Int(l), Int(r)) => Int(l $math_op r),
-                ($math_op_name, Float(l), Float(r)) => Float(l $math_op r),
-            )*
-            // $(
-            //     (BinOp::Equals, $eq_type(l), $eq_type(r)) => Bool(l == r),
-            //     // TODO: Should we say 1 != 1.0, what about "true" != 3
-            //     (BinOp::NotEquals, $eq_type(l), $eq_type(r)) => Bool(l != r),
-            // )*
-            $(
-                ($comarison_name, Int(l), Int(r)) => Bool(l $comparison_op r),
-                ($comarison_name, Float(l), Float(r)) => Bool(l $comparison_op r),
-            )*
-
-            $( $user_lhs => $user_rhs, )*
-
-            (o, l, r) => return Err($env.binop_err(l, o, r, $l_span, $o_span, $r_span).into()),
         }
     };
 }
@@ -305,161 +268,6 @@ impl<'a> Env<'a> {
         }))
     }
 
-    fn binop_err(
-        &self,
-        l: Value,
-        o: BinOp,
-        r: Value,
-        l_span: Span,
-        o_span: Span,
-        r_span: Span,
-    ) -> RtError {
-        RtError(
-            Diagnostic::error()
-                .with_message(format!(
-                    "Unknown binop `{}`, for `{}` and `{}`",
-                    o,
-                    self.type_name(&l),
-                    self.type_name(&r),
-                ))
-                .with_labels(vec![
-                    o_span.primary_label().with_message("In this operator"),
-                    l_span
-                        .secondary_label()
-                        .with_message(format!("LHS evaluated to {:?}", self.dbg_val(&l))),
-                    r_span
-                        .secondary_label()
-                        .with_message(format!("LHS evaluated to {:?}", self.dbg_val(&r))),
-                ]),
-        )
-    }
-
-    fn binop(
-        &mut self,
-        l: Value,
-        o: BinOp,
-        r: Value,
-        l_span: Span,
-        o_span: Span,
-        r_span: Span,
-    ) -> Result<Value> {
-        use Value::*;
-
-        Ok(binop_match!(
-            (o, l, r),
-            (l_span, o_span, r_span, self),
-            {
-                BinOp::Plus   => +,
-                BinOp::Minus  => -,
-                BinOp::Times  => *,
-                BinOp::Devide => /,
-            },
-            // {
-            //     // TODO: What should null == null return
-            //     // TODO: What about Comparisons of non equal types
-            //     /*String,*/ Int, Float, Bool
-            // },
-            {
-                BinOp::GreaterThan       => >,
-                BinOp::GreaterThanEquals => >=,
-                BinOp::LessThan          => <,
-                BinOp::LessThanEquals    => <=,
-            },
-            {
-                // (BinOp::Plus,       String(l), String(r)) => String(l + &r),
-                (BinOp::LogicalOr,  Bool(l),   Bool(r))   => Bool  (l || r),
-                (BinOp::LogicalAnd, Bool(l),   Bool(r))   => Bool  (l && r),
-                (BinOp::Equals, l, r) => {Value::Bool(self.binop_eq(l, o, r, l_span, o_span, r_span)?)},
-                (BinOp::NotEquals, l, r) => {Value::Bool(!self.binop_eq(l, o, r, l_span, o_span, r_span)?)},
-
-                (op, Complex(lid), Complex(rid)) => self.complex_binop(
-                    lid,
-                    op,
-                    rid,
-                    l_span,
-                    o_span,
-                    r_span,
-                )?,
-            },
-        ))
-    }
-
-    fn binop_eq(
-        &self,
-        l: Value,
-        o: BinOp,
-        r: Value,
-        l_span: Span,
-        o_span: Span,
-        r_span: Span,
-    ) -> Result<bool> {
-        Ok(match (l, r) {
-            (Value::Int(l), Value::Int(r)) => l == r,
-            (Value::Float(l), Value::Float(r)) => l == r,
-            (Value::Bool(l), Value::Bool(r)) => l == r,
-            (Value::Complex(lid), Value::Complex(rid)) => {
-                match (&self.heap[lid], &self.heap[rid]) {
-                    (BigValue::String(l), BigValue::String(r)) => l == r,
-                    (BigValue::Array(ls), BigValue::Array(rs)) => {
-                        if ls.len() != rs.len() {
-                            false
-                        } else {
-                            ls.iter().zip(rs).try_fold(true, |acc, (l, r)| {
-                                Result::<_, eyre::Error>::Ok(
-                                    acc && self.binop_eq(*l, o, *r, l_span, o_span, r_span)?,
-                                )
-                            })?
-                        }
-                    }
-                    _ => Err(self.binop_err(
-                        Value::Complex(lid),
-                        o,
-                        Value::Complex(rid),
-                        l_span,
-                        o_span,
-                        r_span,
-                    ))?,
-                }
-            }
-            (Value::Null, Value::Null) => true,
-            (l, r) => Err(self.binop_err(l, o, r, l_span, o_span, r_span))?,
-        })
-    }
-
-    fn complex_binop(
-        &mut self,
-        lid: HeapKey,
-        o: BinOp,
-        rid: HeapKey,
-        l_span: Span,
-        o_span: Span,
-        r_span: Span,
-    ) -> Result<Value> {
-        Ok(match (o, &self.heap[lid], &self.heap[rid]) {
-            // (BinOp::Equals, _, _) => Value::Bool(self.binop_eq(lid, rid, l_span, o_span, r_span)?),
-            // (BinOp::NotEquals, _, _) => {
-            //     Value::Bool(!self.binop_eq(lid, rid, l_span, o_span, r_span)?)
-            // }
-            (BinOp::Plus, BigValue::String(l), BigValue::String(r)) => {
-                let res = l.to_owned() + r;
-                let key = self.heap.insert(BigValue::String(res));
-                Value::Complex(key)
-            }
-            _ => {
-                return Err(self
-                    .binop_err(
-                        Value::Complex(lid),
-                        o,
-                        Value::Complex(rid),
-                        l_span,
-                        o_span,
-                        r_span,
-                    )
-                    .into())
-            }
-        })
-    }
-
     fn unary_op(&self, o: Spanned<UnaryOp>, v: Value, vs: Span) -> Result<Value> {
         Ok(match (o.node, v) {
             (UnaryOp::Not, Value::Bool(b)) => Value::Bool(!b),
@@ -503,23 +311,9 @@ impl<'a> Env<'a> {
         ValueDbg { v, e: self }
     }
 
-    // TODO: This should return a string
     fn print_value(&self, v: &Value) {
         let dbg = self.dbg_val(v);
         println!("{}", dbg);
-        // match v {
-        //     // String(s) => println!("{}", s),
-        //     Int(x) => println!("{}", x),
-        //     Float(f) => println!("{}", f),
-        //     Null => println!("null"),
-        //     Bool(b) => println!("{}", b),
-        //     Complex(id) => match &self.heap[*id] {
-        //         BigValue::String(x) => {
-        //             println!("{}", x)
-        //         }
-        //         BigValue::Array(_) => println!("TODO: FIXME"),
-        //     },
-        // }
     }
 
     fn type_name(&self, v: &Value) -> &'static str {
